automata_requirements.txt
aiohttp
pywinauto
python-dotenv
psutil
keyboard
pydantic
loguru
pyautogui
